+++ testing/experimental/vector/fill_insert.cpp
@@ -18,7 +18,7 @@ void test_reallocating_fill_insert()
 
     assert(result == v.begin());
     assert(v.size() == num_elements_to_insert);
-    assert(std::count(v.begin(), v.end(), 7) == num_elements_to_insert);
+    assert(std::count(v.begin(), v.end(), 7) == static_cast<int>(num_elements_to_insert));
   }
 
   {
@@ -33,8 +33,8 @@ void test_reallocating_fill_insert()
 
     assert(result == v.begin());
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(result, result + num_elements_to_insert, 7) == num_elements_to_insert);
-    assert(std::count(result + num_elements_to_insert, v.end(), 13) == num_initial_elements);
+    assert(std::count(result, result + num_elements_to_insert, 7) == static_cast<int>(num_elements_to_insert));
+    assert(std::count(result + num_elements_to_insert, v.end(), 13) == static_cast<int>(num_initial_elements));
   }
 
   {
@@ -49,8 +49,8 @@ void test_reallocating_fill_insert()
 
     assert(result == v.end() - num_elements_to_insert);
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(result, result + num_elements_to_insert, 7) == num_elements_to_insert);
-    assert(std::count(v.begin(), result, 13) == num_initial_elements);
+    assert(std::count(result, result + num_elements_to_insert, 7) == static_cast<int>(num_elements_to_insert));
+    assert(std::count(v.begin(), result, 13) == static_cast<int>(num_initial_elements));
   }
 
   {
@@ -66,13 +66,13 @@ void test_reallocating_fill_insert()
     auto insertion_end = insertion_begin + num_elements_to_insert;
 
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(insertion_begin, insertion_end, 7) == num_elements_to_insert);
+    assert(std::count(insertion_begin, insertion_end, 7) == static_cast<int>(num_elements_to_insert));
 
     size_t num_initial_elements_before = insertion_begin - v.begin();
     size_t num_initial_elements_after  = v.end() - insertion_end;
 
-    assert(std::count(v.begin(), insertion_begin, 13) == num_initial_elements_before);
-    assert(std::count(insertion_end, v.end(), 13) == num_initial_elements_after);
+    assert(std::count(v.begin(), insertion_begin, 13) == static_cast<int>(num_initial_elements_before));
+    assert(std::count(insertion_end, v.end(), 13) == static_cast<int>(num_initial_elements_after));
   }
 }
 
@@ -95,8 +95,8 @@ void test_nonreallocating_fill_insert()
 
     assert(result == v.begin());
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(result, result + num_elements_to_insert, 7) == num_elements_to_insert);
-    assert(std::count(result + num_elements_to_insert, v.end(), 13) == num_initial_elements);
+    assert(std::count(result, result + num_elements_to_insert, 7) == static_cast<int>(num_elements_to_insert));
+    assert(std::count(result + num_elements_to_insert, v.end(), 13) == static_cast<int>(num_initial_elements));
   }
   
   {
@@ -113,8 +113,8 @@ void test_nonreallocating_fill_insert()
 
     assert(result == v.end() - num_elements_to_insert);
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(result, result + num_elements_to_insert, 7) == num_elements_to_insert);
-    assert(std::count(v.begin(), result, 13) == num_initial_elements);
+    assert(std::count(result, result + num_elements_to_insert, 7) == static_cast<int>(num_elements_to_insert));
+    assert(std::count(v.begin(), result, 13) == static_cast<int>(num_initial_elements));
   }
 
   {
@@ -132,13 +132,13 @@ void test_nonreallocating_fill_insert()
     auto insertion_end = insertion_begin + num_elements_to_insert;
 
     assert(v.size() == num_initial_elements + num_elements_to_insert);
-    assert(std::count(insertion_begin, insertion_end, 7) == num_elements_to_insert);
+    assert(std::count(insertion_begin, insertion_end, 7) == static_cast<int>(num_elements_to_insert));
 
     size_t num_initial_elements_before = insertion_begin - v.begin();
     size_t num_initial_elements_after  = v.end() - insertion_end;
 
-    assert(std::count(v.begin(), insertion_begin, 13) == num_initial_elements_before);
-    assert(std::count(insertion_end, v.end(), 13) == num_initial_elements_after);
+    assert(std::count(v.begin(), insertion_begin, 13) == static_cast<int>(num_initial_elements_before));
+    assert(std::count(insertion_end, v.end(), 13) == static_cast<int>(num_initial_elements_after));
   }
 }
